+++ tests/datasource/fluent/test_sql_datasources.py
@@ -31,14 +31,6 @@ LOGGER = logging.getLogger(__name__)
 
 
 @pytest.fixture
-def create_engine_spy(mocker: MockerFixture) -> Generator[mock.MagicMock, None, None]:  # noqa: TID251 # FIXME CoP
-    spy = mocker.spy(sa, "create_engine")
-    yield spy
-    if not spy.call_count:
-        LOGGER.warning("SQLAlchemy create_engine was not called")
-
-
-@pytest.fixture
 def gx_sqlalchemy_execution_engine_spy(
     mocker: MockerFixture, monkeypatch: pytest.MonkeyPatch
 ) -> Generator[mock.MagicMock, None, None]:  # noqa: TID251 # FIXME CoP
@@ -53,20 +45,6 @@ def gx_sqlalchemy_execution_engine_spy(
     yield spy
     if not spy.call_count:
         LOGGER.warning("SqlAlchemyExecutionEngine.__init__() was not called")
-
-
-@pytest.fixture
-def create_engine_fake(monkeypatch: pytest.MonkeyPatch) -> Generator[None, None, None]:
-    """Monkeypatch sqlalchemy.create_engine to always return a in-memory sqlite engine."""
-    in_memory_sqlite_engine = sa.create_engine("sqlite:///")
-
-    def _fake_create_engine(*args, **kwargs) -> sa.engine.Engine:
-        LOGGER.info(f"Mock create_engine called with {args=} {kwargs=}")
-        return in_memory_sqlite_engine
-
-    monkeypatch.setattr(sa, "create_engine", _fake_create_engine, raising=True)
-    yield
-    in_memory_sqlite_engine.dispose()
 
 
 @pytest.mark.unit
@@ -110,9 +88,9 @@ def create_engine_fake(monkeypatch: pytest.MonkeyPatch) -> Generator[None, None,
     ],
 )
 class TestConfigPasstrough:
+    @pytest.mark.usefixtures("create_engine_spy")
     def test_kwargs_passed_to_create_engine(
         self,
-        create_engine_spy: mock.MagicMock,  # noqa: TID251 # FIXME CoP
         monkeypatch: pytest.MonkeyPatch,
         ephemeral_context_with_defaults: EphemeralDataContext,
         ds_kwargs: dict,
@@ -125,7 +103,7 @@ class TestConfigPasstrough:
         print(ds)
         ds.test_connection()
 
-        create_engine_spy.assert_called_once_with(
+        sa.create_engine.assert_called_once_with(  # type: ignore[union-attr]
             "sqlite:///",
             **{
                 **ds.dict(include={"kwargs"}, exclude_unset=False)["kwargs"],
